Pins database and notifications changes
Pins are saved in the database when pause pins is pressed from the notification. This because it can be (and will possibly be) the last thing the app does.
Pins are not notified about if there was an error with them.

In the future it should be a retry system, where if a pin failed to load 3 times, indicate it as an error. Also differentiate between 404 and load error in the future.