Various fixes for Android < 5.0
Use the AlertDialog from the support library everywhere.
The toolbar was not the last view in the viewgroup, and since the elevation is ignored on <5.0 it became visible when a collapsing toolbar was used.
The FAB has a wrong margin when the parent CoordinatorLayout has a top margin, to account for non collapsing toolbars. Since this is only an issue on not so important layouts, ignore this issue.