Keep controllers attached but hide them, instead of removing the view
Previously when a controller because hidden, it was removed from the view hierachy, and reattached when shown again.
Now the views are only detached from the view hierachy when the whole controller gets destroyed, and the view visibility is changed on onHide and onShow.
This was done for performance reasons. Reattaching a controller takes too much time and is noticable when trying to swipe a controller away.