Add option to select with network type videos should load on
Still depends on the image network type, so the other options are greyed out where appropiate.
For example when you select the image network type to "wifi only", you then can't select the video network type to "all".
The video mode will be set to a lower type automatically when the image mode is changed to something lower.