Refactor WatchManager
Use structured interval types: a handler for foreground update intervals and the alarmmanager for background update intervals.
Fix issue where the backoff would not work on pin watching.
Start of making all database calls optionally async and in a transaction.
Acquire a short wakelock while loading threads in the background.
Update the notifier with peeking and remove ticker and count support.
Shorten quotes and subject.