Add a better save location picker
Taken from another project, might as well use it here.
This layout did allow switching between storage devices, but I found out
how the storage framework works now, and it's too much work to get it
to work right now, so the selection of storage is hidden. Reading works
fine but writing requires a lot more code. Might pick it up later if
there's a good library that does it for us. All the current code uses
the normal File api so that has to be rewritten afaik.

Closes #202