add fastscroll support.
the scroller is attached to the thread recyclerview. it is a modified
version of the FastScroller recently added to the support library,
that adds support for padding and some other tweaks.

the scrollbar looks just as thin as the normal scrollbar, but the grab
area is bigger than the visible bar.