replace dagger with feather
feather is more lightweight, and does not use an annotation processor.
this speeds up the build.

it does use reflection, but that is no issue for us. our graph is really
small, and reflection is not slow on newer android versions.

replaces all calls to the original getGraph() with inject and related
methods.

also update the licenses.