fix opening an url showing the toolbar from the board view.
the slidingpanelayout doesn't call onPanelOpen/Close when the view
is currently in layout. this happens when it is first opened for the
board view and then the thread view without having a layout done.
this calls the same methods at onPanelOpen/Close to fix it when the view
has not been laid out (no dimensions).