setup flow: delete introduction controllers, new empty layout flow.
instead of immediately guiding the user to the sites setup, show the
empty state first and guide them towards the site setup.

the introduction controllers were a big buggy anyway, and this should
help getting the context of the app right the first time.
also rework some strings to be friendlier and less redundant.