exoplayer: fix audio icon missing after swiping.
the icon has to be reset at the start of onPageSwiped To, because
the explayer callback is immediate and otherwise the reset would
happen after the callback.
also a small change to setVolume to look at the connected players
and not the setting (setting can be changed while playing a video
theoretially).