Fix java.lang.IndexOutOfBoundsException: Inconsistency detected.
It happened because displayList from PostAdapter was mutated on a background thread and then used by the RecyclerView on the main thread.
Hidden post filtering now blocks main thread. It won't throw an exception because the database call is being executed on a background thread but the result is being awaited blockingly (There is a comment in ThreadListLayout why is it implemented like this).
To speed up hidden post filtering three SQL indexes were created for PostHide table.