Various fixes
Removed onDowngrade method. DowngradeNotSupportedException is thrown instead which will notify the user that downgrade is not supported and it is pointless to try to import those settings.
Added @Nullable annotation for ExportedAppSettings fields that may become null after json deserialization.
Updated some error messages
Add comment in DatabaseHelper for other devs to notify them that they need to apply their database changes to ExportedAppSettings as well.