build/net: use upstream Volley
Removes all custom volley code we had with a dependency on the official
volley dependency.
This fixes any build issues with the old apache http library, since that
library doesn't exist on new platforms anymore.
We had custom logic in the DiskBasedCache to limit the amount of cache
files, we'll see if this has any impact now. It used to be a big slow
down at start-up.
The cache size has also gone from 10MB to 5MB because we can't change it.