Change format of the exported_clover_settings
Add file header with magic constant to figure out whether it is the Clover settings file or not.
Move out settings version from the JSON part to the header so that we can check whether we can import this version of the file or not before parsing the json (because it may change between versions like how it has happened in the version 2).