MultiImageView: use FileCache's MediaSource for videos
Allows it to stream videos, instead of needing to download the file
entirely. If the video is already on cache, will use it instead.

There's a bit of code-duplication when creating the ExoPlayer, between
the old and the new code paths, but fixing that will make it more
complex than just leaving it there.