Active download can now be stopped or cancelled.
When stopping a download it just stops it's current execution but it's already downloaded files remain on the disk.
When cancelling a download it's files are deleted from the disk as well.
"Save thread" button is now correctly starts/stops/cancels an active download.
Implement "Save thread" button's icon state change (when not downloading a thread it is grey, when downloading it is white).