Fix spoiler images not being shown in saved threads
Fix an image failing the downloading every time when it's extension differs with thumbnail extension thus making such images and their posts trying to be re downloaded every single time until the thread dies.
When thread dies while it's being downloaded, background watcher is disabled and we are in the thread - update the saved thread info in the DB to avoid it getting stuck in the stopped mode but actually being fully downloaded (this makes the downloading icon being shown as "downloading is in progress" even when the thread has already been fully downloaded).
Auto load videos in saved threads when Wi-Fi is turned off.