ui refresher
replace "No." with "#".
show gray background for postthumbnails, to indicate that it's still
loading (like a placeholder).
round thumbnails more in thread and album view.
animate the rounding from/to the image viewer.
when browsing threads and clicking a thimbnail, hide the image for the
transitions, so it looks like the thumbnail itself is moved, instead of
a "copy".
use theme background for post replies popup.