commentparser: handle internal links with whitelist of domain
The fullQuote and quote mathers are now still run on the path segment
of the given url, if the domain is whitelisted as such.
We do this by stripping off the domain from the href if it has one.
If it ends up not matching quote patterns we still use the href as a
normal link.