Fix pin image loading for deleted threads
When we have a local copy of a thread and the thread have been downloading has been deleted we start loading pin image from the local thread instead. But there were to request to load the pin preview: one is via the old setUrl and the other one via setImageFromDisk. They were both executed even for already deleted threads. So what was happening is we were loading an image from the disk and then, after some time when we receive a response from the server with 404 status code, we would replace that image with 404 image.