Lots of minor fixes related to local threads and performance in general:
Mark pin as not being watched when thread gets 404ed or archived.
Check whether a pin has DownloadPosts flag when a pinned thread gets archived or closed before marking it as downloaded (because we might not been downloading it).
Overload equals and hashcode for pin to safely store them in collections.
Before trying to load a thread from the disk check whether a pin associated with it has DownloadPosts flags (performance).
Update loadable in the database when changing lastLoaded, lastViewed etc. fields.
Some methods that access database have been moved to background threads.
Use cached archivedIcon in DrawerAdapter instead of decoding it every single time.