More local threads fixes:
Switching between live thread and local thread now correctly updates download icon.
Fixed a bug when switching to local thread when live thread is still alive (not archived and not 404ed) would make this thread permanently archived (even though live thread is still alive).
Added more logs in ChanThreadLoader so it is more obvious what is happening with downloading thread.
Disable all clicks to save thread button when view local copy and there is still a live thread because it was causing lots of other bugs.