(#172) Attempt to implement migration from one local threads base directory to another
Basically, when a user want to change a local threads directory to another one (let's say they had it on sd-card and then bought another one and want to use it from now on). This is kinda working but it's fucking hacky and some parts of it are really slow.
Fixed a bug with save thread icon being set to the default state when it should be set to stopped state