Update FSAF to version alpha19
- Apparently some users sometimes lose access to their SAF base directories.
This may actually happen when the SAF base directory URI looses the read/write permissions.
Users can do that in settings.
So now FSAF will check whether the SAF base directory still has both permissions and if not it will log it.
So the next time when somebody is complaining that they have to set the base directory again they would be able to provide logs, and will logs it will become obvious wtf is going on. (It may actually be another reason, but this one is the most possible for now).