A little improvement for swipe-to-save gesture
The previous swipe-to-save detection was not ideal. If you have a wide image (width > height) and you slightly zoom in you might expect that swipe-to-save would be replaced with swipe-to-close gesture but that wasn't true. That's because the viewport would still touch both the top and the bottom sides of the image. So to actually trigger swipe-to-close instead of swipe-to-save you would have to zoom further in until you can start panning the viewport vertically. While testing this thing it turned out to be a little bit irritating. So instead, it is now checking whether all sides of the viewport are touching all sides of an image before triggering swipe-to-save instead of swipe-to-close. So, basically, it is now only allowed when you are fully zoomed out (the default image state when you open it).