(#562) Fix a request not getting canceled when encountering a 404 response from the server
If we encounter an exception in ChunkPersister and the current state is Running, that's find, that just means that an exception other than CancellationException was thrown. So that means that we only need to
do the canceling of the current request, pass the exception downstream and that's it. It will be handled by the downstream.