Closes #1224
This commit does a number of things:
There's now a DrawableRepository, so that one drawable can just be constantly reused
ThumbnailView now only takes one parameter into setUrl, as it is intended for things that should be square in dimension (or at least, both should be the same dimension)
PostAdapter and PostImagesAdapter now properly recycle cells, so no inconsistencies should occur from that
PostImageThumbnailView long clicks now only apply to those in CardPostCell, PostCell, and AlbumViewCell
The preview image (used when going into/out of the ImageViewerController) decodes to the thumbnail exactly now; it doesn't use the full size image. This may change in the future, architecture permitting
The xml attribute decode_dimen has been removed, as we now grab the layout params where applicable, or decode to the full view size

And most importantly, the issue in the title is fixed. This was caused by pre-draw listeners not being removed in time (or at all), so now they are removed swiftly when applicable and should not cause any more issues